requirements.txt
numpy==1.24.3
Pillow==11.1.0
